[next-lint] test: remove eslint config snapshot testing (#77818)
### Why?

The snapshot testing occasionally failed when the rules were modified in
newer ESLint versions. This is because we snapshot-tested the whole
eslint config output. However, this snapshot test is intended to verify
if the necessary plugins were loaded, which can be done by checking the
`plugins` property.

### How?

Removed snapshot testing and added `plugins` to be checked along with
the existing `parser` and `settings` checking.