[dev-overlay] sync horizontal scrollbar style (#77769)
Following up on https://github.com/vercel/next.js/pull/76953.

In macOS, the scrollbar color is still light in dark mode. So, we
modified the style for it at
https://github.com/vercel/next.js/pull/76509. However, it didn't cover
the scrollbars of the children of the dialog. Therefore, we sync the
scrollbar of children to be dark in dark mode as well.

| Before | After |
|--------|--------|
| ![CleanShot 2025-04-03 at 09 59
01@2x](https://github.com/user-attachments/assets/244566ad-4c05-4045-891f-eff8a7acda17)
| ![CleanShot 2025-04-03 at 09 57
18@2x](https://github.com/user-attachments/assets/6441251b-eb0d-450d-8a32-75d9a4d88f32)
|