[next-server] skip setting vary header for basic routes (#77797)
### What

After https://github.com/vercel/vercel/pull/13011 landed on Vercel CLI,
noticing that we're still checking the `vary` header in the tests, which
is obselete now. In this PR we delete the handling for setting the
`vary` header but still keep the cases of interception routes since it's
still needed.

### Why

Since v13.4.6 we added the `_rsc` query with the unique hash, `vary`
header becomes not required anymore.

Closes NDX-1010