[Experiment] <Link unstable_dynamicOnHover>
This adds an unstable option to the Link component's called
`unstable_dynamicOnHover`. When set to true, the Link will initiate a
dynamic prefetch as soon as the link is hovered.

This is roughly equivalent to updating the prefetch prop to `true`
inside a mouse event handler.

Because this results in dynamic prefetching, semantically it has the
same implication as `prefetch={true}` — by the time the navigation
occurs, the target page's data may be slightly stale.