[ts-next-plugin] fix: use `getSourceFile` instead of `fileExists` to check file existence (#77863)
### Why?

The Next.js TypeScript plugin was doing a workaround for updating the
virtual file with `deleteFile()` then `createFile()` since
`updateFile()` failed even though `env.sys.fileExists()` passed.

It is because the system was created based on the FS system via
`createFSBackedSystem()`, which falls back to read from the actual file
system if the file doesn't exist on the current virtual file system.


https://github.com/microsoft/TypeScript-Website/blob/9d447358e4fc67c881202fb40115cd8adf70daad/packages/typescript-vfs/src/index.ts#L544

![CleanShot 2025-04-05 at 17 39
35@2x](https://github.com/user-attachments/assets/96b3ec2e-3b6d-407d-877f-bdd3224b156b)

### How?

Use `getSourceFile()` to check if the file actually exists. This is
what`@typescript/vfs` uses to check before running the `updateFile()`.


https://github.com/microsoft/TypeScript-Website/blob/9d447358e4fc67c881202fb40115cd8adf70daad/packages/typescript-vfs/src/index.ts#L83-L87