Add experimental.routerBFCache to NextConfig
Sets up a flag that will depend on an experimental React feature
(Activity). I based this on the existing code for the
`viewTransition` flag.

I haven't implemented anything behind this flag yet; this only wires up
the flag itself. I added a basic test that renders the Activity
component to verify that it's available when the flag is on.