test: attempt to unflake useLinkStatus interrupted navigation test (#77969)
This test flaked for me here:
https://github.com/vercel/next.js/actions/runs/14353556171/job/40239845657#step:33:678

I suspect this is because \[the assertion that the pending state should
disappear\] sometimes runs too soon. I made it wait until the
side-effect of clicking the "debug mode" button becomes visible before
checking if the pending state is gone. If this fix doesn't improve the
situation, it indicates a bug in `useLinkStatus`, because the pending
state is still hanging around for some reason.

(i chose not to just wrap the assertion in a `retry()`, because we want
to ensure that the indicator disappears immediately, not at some
unspecified point in the future)