Revert standalone extension change for module type (#77978)
We don't need to change the extension as the main fix is to ensure the
`package.json` is copied to the standalone folder. The extension change
was for more correctness but can cause breakage for anyone using `type:
'module'` as they will need to update their flows for the new `.mjs`
extension. Since this is un-necessary breakage this reverts that piece
of the fix.

x-ref:
https://github.com/vercel/next.js/pull/77944#discussion_r2035618277