Uninstall uncaughtException listeners, too
This broadens the scope of the `removeUnhandledRejectionListeners`
flag to also remove `uncaughtException` listeners.

The motivation is the same: uncaught exceptions in one part of the
page should not prevent the rest of the page from rendering.

See https://github.com/vercel/next.js/pull/77997 for more context.

I renamed the experimental flag to `removeUncaughtErrorAndRejectionListeners`
to better reflect increased scope but the name isn't really that
important since at this stage we just need to try out the behavior; it
isn't documented anywhere.