test: adjust default timeouts for [waitFor]elementByCss (#78026)
Adjusts the default timeouts:
- elementByCss: 5s
- waitForElementByCss: 10s

Previously, we didn't pass an explicit value, so they both defaulted to
playwright's default of 60s. This is _very_ slow, and if something is
slow enough to require anything in that range, i'd say it deserves to be
wrapped in `retry()` or something similar anyway, so it's fine to fail
here.
The upside of this is that errors because of bad selectors (or missing
elements) happen faster, which is a nice thing, and might also save CI
time.

i could also see an argument for making these even shorter -- e.g. the
default timeout for `retry()` is 3s, so maybe `retry` should be faster
than that.

https://github.com/vercel/next.js/blob/717e54b0cd914a17314c470cd10d6af2b2d75787/test/lib/next-test-utils.ts#L796-L799
but then again, `retry` doesn't factor in the time `fn` took, so these
don't really conflict.