[dynamicIO] use flight prerender even with stable React
When dynamicIO was first implemented the prerender API for flight was not available in the stable channel of React. We use a different method when dynamicIO is enabled without PPR than with PPR b/c we did not want to requrie experimental React if you weren't using an experimental only API.

Now that prerender is available in the stable channel we can adopt it regardless of channel. This change updates the implementation to use this API.

The main difference in these approaches is `prerender` will leave incomplete chunks in the prerendered RSC stream unresolevd when you abort. This is useful for future cases where we will support resuming in RSC. It is also better models the idea that errors can be retried in an environment closer to the user. For instance when SSR errors we don't trigger ErrorBoundaries on the server we instead put the nearest Suspense into fallback mode and have React in the browser try to recover. Similarly if prerender completes with an error we don't need to encode those errors into the stream but can instead expect the consumer to attempt to recover. This might look like a resume render during an MPA nav or a full navigation render during an SPA nav.

One consequence of this is there will be no PRERENDER_INTERRUPTED errors encoded into the RSC stream. This can slightly degrade the error specificity when you violate the rules of dynamicIO during builds but this was already the case if you enabled PPR and that is where this is all heading anyway. We have some additional plans on how we might pinpoint the exact cause of invalid dynamic using a diagnostic prerender but that will be added in a future update.