Turbopack: fix cell not found bug (#78246)
### What?

* Some debug asserts were missing to validate the correct category on read access via `iter()` or `extract_if()`. Added these.
* incorrect accessed category led to empty list of dependencies after restoring from persistent caching, which led to weird bugs like Cell not found errors

So optimized the category assignment to make sure aggregation updates can operate on Meta only. Children, Collectibles and CollectiblesDependent were moved to Meta category since they are accessed during aggregation updates. `Children` are used as follower for non-aggregating nodes. `Collectibles` are read when computing the aggregated information and `CollectiblesDependent` are read when aggregated collectibles are modified.

* Added a unit test reproducting this bug and the next PR in the stack also added a e2e test reproducting it.