fix: amphtml-validator WASM errors (for real) (#78379)
Rejoice! The day has come. Hopefully, we'll never see `AssertionError:
Assertion failed: WebAssembly is uninitialized` ever again. I dug into
amphtml-validator's code, and discovered that they have a race condition
when initializing. This PR works around it.

The root problem is that `AmpValidator.getInstance` is buggy when
multiple calls to it occur in parallel.
See, the point of `AmpValidator.getInstance(validatorPath)` is to re-use
instances across multiple calls. It attempts to do that by stashing
created instances in `instanceByValidatorJs`.
However, when creating a fresh instance, it is added to
`instanceByValidatorJs` **before `instance.init()` is finished**:

https://github.com/ampproject/amphtml/blob/0c8eaba73ca8f5c462a642fa91901a29e6304f6e/validator/js/nodejs/index.js#L309-L313

which means that, if a concurrent call to
`AmpValidator.getInstance(validatorPath)` happens before the original
`instance.init()` is done, then we'll get back an uninitialized or
partially initialized instance:

https://github.com/ampproject/amphtml/blob/0c8eaba73ca8f5c462a642fa91901a29e6304f6e/validator/js/nodejs/index.js#L292-L294

And, since `instance.init()` is the part that handles loading the
webassembly code, this race results in the dreaded `AssertionError:
Assertion failed: WebAssembly is uninitialized` error when calling
methods on the not-yet-initialized validator.

As a workaround, we properly dedupe instance creation on our own. I'll
submit a fix to the original repo as well, but this should solve the
problem for us.