fix(turbo-tasks): Add a few more transient-from-persistent task assertions (#78301)
Transient tasks are tasks that depend on a value that cannot be serialized. Transient tasks and anything that depend on them therefore cannot be serialized. Because of this, it's not valid for persistent (serialized) tasks to depend on transient tasks.

We're seeing issues from transient collectibles being read from persistent tasks.

This should only be possible if a persistent task emits a transient collectible. That shouldn't be possible either (where would it get the transient collectible from?), but asserting doesn't seem like a bad idea, and might get us a stack trace closer to the root-cause.