test: attempt to de-flake graceful shutdown test (#78464)
the test has been flaking recently because (i guess) it takes a bit
longer for the app to start rejecting new requests than the test was
expecting. i wrapped the assertion in a `retry` so that the test is
resillient to that.

i've applied the same changes to a previously skipped test and
un-skipped it -- i'm hoping it's gonna be okay now. also removed some
unnecessary assertions about the app still being active at a given
point, we don't really care about that.