feat(turbopack): Implement side-effect optimization (#78047)
### What?

Implement side-effect-free optimization for turbopack. This is required to build applications properly and to reduce the final bundle size. The advanced tree shaking consists of module splitting and module merging. The module splitting is done only once per module, and module merging is done naturally by the chunking functions because split modules are still modules. This PR implements optimization for the module merging part - the optimization in this PR allows skipping intermediate modules between the requested module and the resolved module.



### Why?



### How?
 - Closes PACK-3500

 - Closes #72947
 	- This is a retry of https://github.com/vercel/next.js/pull/72947.

---------

Co-authored-by: Niklas Mischkulnig <4586894+mischnic@users.noreply.github.com>