fix(turbo-tasks): Store persistence of wrapped task on RawVc::LocalOutput (#78488)
Using the parent task id for determining the persistence of a `RawVc::LocalOutput` is (sometimes) wrong.

If we have a transient task calling a persistent task, the wrapped `LocalOutput` should be persistent.

We just store the persistence here, and not the task id of the wrapped `LocalOutput`, because when this is constructed, no non-local task id has been reserved yet. We don't really need the full task id though: just the persistence/transient bit is sufficient.

After this PR stack, the parent task id isn't really used for anything anymore, so we can get rid of it (TODO).

Tested using the repro here: https://vercel.slack.com/archives/C046HAU4H7F/p1744792143917369?thread_ts=1744792143.917369