Turbopack: Allow a custom list of file extensions to use with `@next/mdx`
This allows users to customize the file extensions to process with `@next/mdx`.

Before this change, `@next/mdx` with Turbopack did not consider the `extension` property at all. This was challenging since it doesn’t support all of the different ways of testing module specifiers that webpack does, such as with a regular expression.

This implements (and documents) a different behavior for Turbopack: accepting a list of extensions as `extension`. This is fairly backwards compatible, as it does not change how the plugin behaves with webpack, and it was ignored by Turbopack previously.

For example, to configure `next.config.js` with `@next/mdx` and support for both the `mdx` and `md` extension:

```js
import createMDX from '@next/mdx'

const withMDX = createMDX({
  extension: ['mdx', 'md'],
})

export default withMDX({})
```

This is equivalent to the following with Next.js and webpack:

```js
import createMDX from '@next/mdx'

const withMDX = createMDX({
  extension: /\.mdx?$/,
})

export default withMDX({})
```

Test Plan: Changed the existing mdx test to import an `md` component as well.